Switch to linux for docdeploy, push preview (#12)
There seems to be an error in the last step of docdeploy that looks
like a possible windows path error. This was using windows because
at one point it wasn't easy to use xvfb on linux with the docdeploy
action, but that has long since been fixed.

This also pushes a preview and fixes doctest failures.