port to Gtk4 (#53)
* use Gtk4

* support "focus-leave" event for textbox, restore some tests

* fix some tests

* add XML file (formerly player.glade)

* restore commented out code, fix a crash in slider

* fix slider

Gtk4 only rounds when the value is shown, which is not the default

* fix zoom-scroll at mouse position

* work on precompile code

* fix a few doctest problems, remove 1.3 from CI

* fix the canvas precompiles that I know how to fix

* don't open windows during precompilation

* bump Gtk4 compat

* implement rest of tests

This introduces a way to override the modifier state for the purposes of testing. It's
better than nothing.

* add precompile code for events with modifiers

Also disable player precompile code, which freezes in Julia 1.10 due to a Timer not closing. This will be fixed in a future commit.