fixes for ImageView port to Gtk4 (#54)
Fixes issues discovered while polishing the port of ImageView to Gtk4.

First, the mouse scrolling direction was inadvertently reversed. I'm not sure if this is a change in the convention of GTK4 vs. GTK3 or a mistake I made. In any case this PR restores the original direction.

Second, this tweaks the GtkBuilder XML for the player widget.

Third, I added a filter to remove zoom events (actually, those with a CONTROL modifier, which is the default for zoom) to avoid panning while zooming.